fix unsupported rapidsjon value type

+++ src/tyr/locate_serializer.cc
@@ -151,7 +151,7 @@ void serialize_edges(const PathLocation& path_location,
         writer.start_object("bounding_circle");
         writer("lat", edge.bounding_circle.first.lat());
         writer("lon", edge.bounding_circle.first.lng());
-        writer("radius", edge.bounding_circle.second);
+        writer("radius", static_cast<uint64_t>(edge.bounding_circle.second));
         writer.end_object();
 
         writer("linear_reference", linear_reference(directed_edge, edge.percent_along, edge_info));
